Introducing control flow analysis
Control flow proceeds as a directed graph of points. Each point will,
based on the information that is known about the state of the program,
produce further possible states. A conditional construct such as an if
statement will, for example, lead to two possible program states, one
where the condition is true and one where the condition is false.

By generating all of the possible control flow states given a known
starting state, a residual program may be generated.

Introduce an analyzer that examines Go statements and transforms them
into points. Provide a library of points.

There have been some other changes to the evaluation/reduction of
expressions, but a lot of that is probably going to be reverted soon as
I'm not happy with references (lvalues) and values sharing the same
objects.