Made binding time analysis more accessible
Move the code for binding time analysis out of the test file and create
an exported API for the rest of the specialiser to use.

This introduces a graph type that can be generated from a control flow
graph specified in terms of an interface that the client implements. The
idea is that you generate the graph once, and then use it many times as
the subprogram that it refers to is specialised during the partial
evaluation process.

This can currently only deal with relatively simple control flow graphs,
although it can handle loops it may need some tweaking to deal with
function application.